micro-services/requirements.txt
feedparser
numpy
transformers
torch
python-telegram-bot
flair
sentencepiece
protobuf
sqlite3